VolleyGson-Xampp

+++ app/src/main/java/com/ahmadabuhasan/volleygsonxampp/RecyclerViewAdapter.java
@@ -3,11 +3,12 @@ package com.ahmadabuhasan.volleygsonxampp;
 import android.annotation.SuppressLint;
 import android.app.AlertDialog;
 import android.content.Context;
-import android.content.DialogInterface;
 import android.content.Intent;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.Filter;
+import android.widget.Filterable;
 import android.widget.TextView;
 import android.widget.Toast;
 
@@ -19,10 +20,8 @@ import com.android.volley.NetworkError;
 import com.android.volley.NoConnectionError;
 import com.android.volley.ParseError;
 import com.android.volley.Request;
-import com.android.volley.Response;
 import com.android.volley.ServerError;
 import com.android.volley.TimeoutError;
-import com.android.volley.VolleyError;
 import com.android.volley.toolbox.StringRequest;
 import com.google.gson.Gson;
 
@@ -30,17 +29,20 @@ import java.util.ArrayList;
 
 /*
  * Created by Ahmad Abu Hasan on 29/12/2020
+ * Update: Search on 07/01/2021
  */
 
-public class RecyclerViewAdapter extends RecyclerView.Adapter<RecyclerViewAdapter.ViewHolder> {
+public class RecyclerViewAdapter extends RecyclerView.Adapter<RecyclerViewAdapter.ViewHolder> implements Filterable {
 
-    private Context context;
-    private ArrayList<ModelBarang> arrayModelBarangs;
+    private final Context context;
+    ArrayList<ModelBarang> arrayModelBarangs, dataFilter;
+    CustomFilter filter;
 
     // membuat kontruksi recyclerviewadapter
     public RecyclerViewAdapter(Context context, ArrayList<ModelBarang> arrayModelBarangs) {
         this.context = context;
         this.arrayModelBarangs = arrayModelBarangs;
+        this.dataFilter = arrayModelBarangs;
     }
 
     @NonNull
@@ -69,11 +71,21 @@ public class RecyclerViewAdapter extends RecyclerView.Adapter<RecyclerViewAdapte
         return arrayModelBarangs.size();
     }
 
+    @Override
+    public Filter getFilter() {
+        if (filter == null) {
+            filter = new CustomFilter(dataFilter, this);
+        }
+        return filter;
+    }
+
     // membuat class viewholder
     public class ViewHolder extends RecyclerView.ViewHolder {
 
         // implementasi textview
-        private TextView textView_nim, textView_name, textView_class;
+        private final TextView textView_nim;
+        private final TextView textView_name;
+        private final TextView textView_class;
 
         public ViewHolder(@NonNull View itemView) {
             super(itemView);
@@ -84,33 +96,27 @@ public class RecyclerViewAdapter extends RecyclerView.Adapter<RecyclerViewAdapte
             textView_class = itemView.findViewById(R.id.tv_class);
 
             // mendeklarasi item ketika diklik
-            itemView.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View v) {
-                    // mendapatkan posisi pada adapter
-                    int position = getAdapterPosition();
-                    // mengambil posisi pada arraymodelbarang
-                    ModelBarang modelBarang = arrayModelBarangs.get(position);
+            itemView.setOnClickListener(v -> {
+                // mendapatkan posisi pada adapter
+                int position = getAdapterPosition();
+                // mengambil posisi pada arraymodelbarang
+                ModelBarang modelBarang = arrayModelBarangs.get(position);
 
-                    String[] pilihan = {"Lihat", "Ubah", "Hapus"};
-                    // menamplkan pilihan alertdialog
-                    new AlertDialog.Builder(context)
-                            .setTitle("Pilihan")
-                            .setItems(pilihan, new DialogInterface.OnClickListener() {
-                                @Override
-                                public void onClick(DialogInterface dialog, int which) {
-                                    if (which == 0) { // 0 sama dengan Lihat
-                                        lihatDataBarang(modelBarang);
-                                    } else if (which == 1) { // 1 sama dengan Ubah
-                                        ubahDataBarang(modelBarang);
-                                    } else if (which == 2) { // 2 sama dengan Hapus
-                                        hapusDataBarang(position, modelBarang);
-                                    }
-                                }
-                            })
-                            .create()
-                            .show();
-                }
+                String[] pilihan = {"Lihat", "Ubah", "Hapus"};
+                // menamplkan pilihan alertdialog
+                new AlertDialog.Builder(context)
+                        .setTitle("Pilihan")
+                        .setItems(pilihan, (dialog, which) -> {
+                            if (which == 0) { // 0 sama dengan Lihat
+                                lihatDataBarang(modelBarang);
+                            } else if (which == 1) { // 1 sama dengan Ubah
+                                ubahDataBarang(modelBarang);
+                            } else if (which == 2) { // 2 sama dengan Hapus
+                                hapusDataBarang(position, modelBarang);
+                            }
+                        })
+                        .create()
+                        .show();
             });
         }
 
@@ -126,12 +132,7 @@ public class RecyclerViewAdapter extends RecyclerView.Adapter<RecyclerViewAdapte
             new AlertDialog.Builder(context)
                     .setTitle("Details")
                     .setMessage(deskripsi)
-                    .setPositiveButton("Ok!", new DialogInterface.OnClickListener() {
-                        @Override
-                        public void onClick(DialogInterface dialog, int which) {
-                            dialog.dismiss();
-                        }
-                    })
+                    .setPositiveButton("Ok!", (dialog, which) -> dialog.dismiss())
                     .create()
                     .show();
         }
@@ -149,56 +150,50 @@ public class RecyclerViewAdapter extends RecyclerView.Adapter<RecyclerViewAdapte
 
         // method untuk menghapus data barang
         private void hapusDataBarang(int position, @NonNull ModelBarang modelBarang) {
-            String url = "http://ipaddressServer/volleygsonxampp/delete?id=" + modelBarang.getMhs_id();
+            String url = ApiClient.URL_DELETE + modelBarang.getMhs_id();
             // buat StringRequest volley dan jangan lupa requestnya GET "Request.Method.GET"
-            StringRequest request = new StringRequest(Request.Method.GET, url, new Response.Listener<String>() {
-                @Override
-                public void onResponse(String response) {
-                    // menerapkan ke model class menggunakan GSON
-                    // mengkonversi JSON ke java object
-                    ResponStatus responStatus = new Gson().fromJson(response, ResponStatus.class);
-                    int status_kode = responStatus.getStatus_kode();
-                    String status_pesan = responStatus.getStatus_pesan();
+            StringRequest request = new StringRequest(Request.Method.GET, url, response -> {
+                // menerapkan ke model class menggunakan GSON
+                // mengkonversi JSON ke java object
+                ResponStatus responStatus = new Gson().fromJson(response, ResponStatus.class);
+                int status_kode = responStatus.getStatus_kode();
+                String status_pesan = responStatus.getStatus_pesan();
 
-                    // validasi status kode
-                    if (status_kode == 1) {
-                        // menampilkan toast berhasil
-                        Toast.makeText(context, status_pesan, Toast.LENGTH_SHORT).show();
+                // validasi status kode
+                if (status_kode == 1) {
+                    // menampilkan toast berhasil
+                    Toast.makeText(context, status_pesan, Toast.LENGTH_SHORT).show();
 
-                        // menghapus pada tampilan arraylist!
-                        notifyItemRemoved(position);
-                        arrayModelBarangs.remove(position);
-                    } else if (status_kode == 2) {
-                        Toast.makeText(context, status_pesan, Toast.LENGTH_SHORT).show();
-                    } else if (status_kode == 3) {
-                        Toast.makeText(context, status_pesan, Toast.LENGTH_SHORT).show();
-                    } else if (status_kode == 4) {
-                        Toast.makeText(context, status_pesan, Toast.LENGTH_SHORT).show();
-                    } else if (status_kode == 5) {
-                        Toast.makeText(context, status_pesan, Toast.LENGTH_SHORT).show();
-                    } else {
-                        Toast.makeText(context, "Status Kesalahan Tidak Diketahui!", Toast.LENGTH_SHORT).show();
-                    }
+                    // menghapus pada tampilan arraylist!
+                    notifyItemRemoved(position);
+                    arrayModelBarangs.remove(position);
+                } else if (status_kode == 2) {
+                    Toast.makeText(context, status_pesan, Toast.LENGTH_SHORT).show();
+                } else if (status_kode == 3) {
+                    Toast.makeText(context, status_pesan, Toast.LENGTH_SHORT).show();
+                } else if (status_kode == 4) {
+                    Toast.makeText(context, status_pesan, Toast.LENGTH_SHORT).show();
+                } else if (status_kode == 5) {
+                    Toast.makeText(context, status_pesan, Toast.LENGTH_SHORT).show();
+                } else {
+                    Toast.makeText(context, "Status Kesalahan Tidak Diketahui!", Toast.LENGTH_SHORT).show();
                 }
-            }, new Response.ErrorListener() {
-                @Override
-                public void onErrorResponse(VolleyError error) {
-                    // jika respon tidak ditemukan maka akan menampilkan berbagai error berikut ini
-                    if (error instanceof TimeoutError) {
-                        Toast.makeText(context, "Network TimeoutError", Toast.LENGTH_SHORT).show();
-                    } else if (error instanceof NoConnectionError) {
-                        Toast.makeText(context, "Nerwork NoConnectionError", Toast.LENGTH_SHORT).show();
-                    } else if (error instanceof AuthFailureError) {
-                        Toast.makeText(context, "Network AuthFailureError", Toast.LENGTH_SHORT).show();
-                    } else if (error instanceof ServerError) {
-                        Toast.makeText(context, "Server Error", Toast.LENGTH_SHORT).show();
-                    } else if (error instanceof NetworkError) {
-                        Toast.makeText(context, "Network Error", Toast.LENGTH_SHORT).show();
-                    } else if (error instanceof ParseError) {
-                        Toast.makeText(context, "Parse Error", Toast.LENGTH_SHORT).show();
-                    } else {
-                        Toast.makeText(context, "Status Error Tidak Diketahui!", Toast.LENGTH_SHORT).show();
-                    }
+            }, error -> {
+                // jika respon tidak ditemukan maka akan menampilkan berbagai error berikut ini
+                if (error instanceof TimeoutError) {
+                    Toast.makeText(context, "Network TimeoutError", Toast.LENGTH_SHORT).show();
+                } else if (error instanceof NoConnectionError) {
+                    Toast.makeText(context, "Nerwork NoConnectionError", Toast.LENGTH_SHORT).show();
+                } else if (error instanceof AuthFailureError) {
+                    Toast.makeText(context, "Network AuthFailureError", Toast.LENGTH_SHORT).show();
+                } else if (error instanceof ServerError) {
+                    Toast.makeText(context, "Server Error", Toast.LENGTH_SHORT).show();
+                } else if (error instanceof NetworkError) {
+                    Toast.makeText(context, "Network Error", Toast.LENGTH_SHORT).show();
+                } else if (error instanceof ParseError) {
+                    Toast.makeText(context, "Parse Error", Toast.LENGTH_SHORT).show();
+                } else {
+                    Toast.makeText(context, "Status Error Tidak Diketahui!", Toast.LENGTH_SHORT).show();
                 }
             });
